requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
google-search-results>=2.4.2
langgraph>=0.0.15
pydantic>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
cachetools>=5.3.2
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.2.0
fpdf>=1.7.2 